rename integration to acceptance tests and remove unit tests
