+++ .github/workflows/acceptance_test.yaml
@@ -1,4 +1,4 @@
-name: Integration Test
+name: Acceptance Tests
 
 on:
   push:
