+++ tests/acceptance/acceptance_suite_test.go
@@ -1,4 +1,4 @@
-package integration
+package acceptance_test
 
 import (
 	"bytes"
@@ -9,39 +9,98 @@ import (
 	"os/exec"
 	"path/filepath"
 	"regexp"
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+	"github.com/onsi/gomega/gexec"
+	"k8s.io/apimachinery/pkg/runtime"
+	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	helmv2 "github.com/fluxcd/helm-controller/api/v2beta1"
+	kustomizev1 "github.com/fluxcd/kustomize-controller/api/v1beta1"
+	sourcev1 "github.com/fluxcd/source-controller/api/v1beta1"
+	profilesv1 "github.com/weaveworks/profiles/api/v1alpha1"
 )
 
 var (
-	waitForDeploymentArgs = []string{
-		"-n",
-		"profiles-system",
-		"wait",
-		"--for=condition=available",
-		"deployment",
-		"profiles-controller-manager",
-		"--timeout",
-		"5m",
-	}
-	waitForPodsArgs = []string{
-		"-n",
-		"profiles-system",
-		"wait",
-		"--for=condition=Ready",
-		"--all",
-		"pods",
-		"--timeout",
-		"5m",
-	}
-	applySourceCatalogArgs = []string{
-		"apply",
-		"-f",
-		"catalog-source.yaml",
-	}
+	binaryPath                      string
+	skipTestsThatRequireCredentials bool
 )
 
-// PrepareTestCluster will create a test cluster using pctl `prepare` command.
+func TestAcceptance(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Acceptance Suite")
+}
+
+var (
+	kClient client.Client
+)
+
+var _ = BeforeSuite(func() {
+	var err error
+	binaryPath, err = gexec.Build("github.com/weaveworks/pctl/cmd/pctl")
+	Expect(err).NotTo(HaveOccurred())
+
+	// overwrite the default test repository location if set
+	if v := os.Getenv("PCTL_TEST_REPOSITORY_URL"); v != "" {
+		pctlTestRepositoryName = v
+	}
+
+	if v := os.Getenv("SKIP_CREDENTIAL_TESTS"); v == "true" {
+		skipTestsThatRequireCredentials = true
+	}
+
+	scheme := runtime.NewScheme()
+	Expect(clientgoscheme.AddToScheme(scheme)).To(Succeed())
+	Expect(profilesv1.AddToScheme(scheme)).To(Succeed())
+	Expect(helmv2.AddToScheme(scheme)).To(Succeed())
+	Expect(kustomizev1.AddToScheme(scheme)).To(Succeed())
+	Expect(sourcev1.AddToScheme(scheme)).To(Succeed())
+
+	kubeconfig := ctrl.GetConfigOrDie()
+	kClient, err = client.New(kubeconfig, client.Options{Scheme: scheme})
+	Expect(err).NotTo(HaveOccurred())
+	Expect(prepareTestCluster(binaryPath)).To(Succeed())
+})
+
+var _ = AfterSuite(func() {
+	gexec.CleanupBuildArtifacts()
+})
+
+// prepareTestCluster will create a test cluster using pctl `prepare` command.
 // @binary -- location of the built pctl binary.
-func PrepareTestCluster(binaryPath string) error {
+func prepareTestCluster(binaryPath string) error {
+	var (
+		waitForDeploymentArgs = []string{
+			"-n",
+			"profiles-system",
+			"wait",
+			"--for=condition=available",
+			"deployment",
+			"profiles-controller-manager",
+			"--timeout",
+			"5m",
+		}
+		waitForPodsArgs = []string{
+			"-n",
+			"profiles-system",
+			"wait",
+			"--for=condition=Ready",
+			"--all",
+			"pods",
+			"--timeout",
+			"5m",
+		}
+		applySourceCatalogArgs = []string{
+			"apply",
+			"-f",
+			"catalog-source.yaml",
+		}
+	)
+
 	tmp, err := ioutil.TempDir("", "prepare_integration_test_01")
 	if err != nil {
 		return fmt.Errorf("failed to create temp folder for test: %w", err)
